+++ backend/src/main/java/com/ssp/assigmnents/sizecalculator/domain/Brand.java
@@ -1,4 +1,4 @@
-package com.ssp.assigmnents.domain;
+package com.ssp.assigmnents.sizecalculator.domain;
 
 public class Brand {
