+++ backend/src/main/java/com/ssp/assigmnents/sizecalculator/domain/DomainFactory.java
@@ -1,4 +1,4 @@
-package com.ssp.assigmnents.domain;
+package com.ssp.assigmnents.sizecalculator.domain;
 
 import java.util.Collection;
 import java.util.stream.Collectors;
